read data pretty much works

+++ student-distrib/kernel.c
@@ -61,7 +61,7 @@ void entry(unsigned long magic, unsigned long addr) {
         module_t* mod = (module_t*)mbi->mods_addr;
 
         //GETTING BOOTBLOCK ADDR
-        get_bootblock_address((unsigned long)mod->mod_start);
+        get_bootblock_address((uint32_t)mod->mod_start);
 
         while (mod_count < mbi->mods_count) {
             printf("Module %d loaded at address: 0x%#x\n", mod_count, (unsigned int)mod->mod_start);
